+++ packages/google_maps_flutter/google_maps_flutter_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.3
+
+* Adds support for heatmap layers.
+
 ## 0.3.2+1
 
 * Removes dependency on `meta`.